Add getFeatureInfoUrl and getLegendUrl functions to ol/wms
