+++ src/ol/source/wms.js
@@ -6,16 +6,22 @@ import {DECIMALS} from './common.js';
 import {appendParams} from '../uri.js';
 import {compareVersions} from '../string.js';
 import {decode} from '../Image.js';
-import {getHeight, getWidth} from '../extent.js';
+import {floor, round} from '../math.js';
+import {getForViewAndSize, getHeight, getWidth} from '../extent.js';
 import {get as getProjection} from '../proj.js';
 import {getRequestExtent} from './Image.js';
-import {round} from '../math.js';
 
 /**
  * Default WMS version.
  * @type {string}
  */
 export const DEFAULT_VERSION = '1.3.0';
+
+/**
+ * @const
+ * @type {import("../size.js").Size}
+ */
+const GETFEATUREINFO_IMAGE_SIZE = [101, 101];
 
 /**
  * @api
@@ -184,4 +190,99 @@ export function createLoader(options) {
     }
     return load(image, src).then((image) => ({image, extent, pixelRatio}));
   };
+}
+
+/**
+ * Get the GetFeatureInfo URL for the passed coordinate and resolution. Returns `undefined` if the
+ * GetFeatureInfo URL cannot be constructed.
+ * @param {LoaderOptions} options Options passed the `createWMSLoader()` function. In addition to
+ * the params required by the loader, `INFO_FORMAT` should be specified, it defaults to
+ * `application/json`. If `QUERY_LAYERS` is not provided, then the layers specified in the `LAYERS`
+ * parameter will be used.
+ * @param {import("../coordinate.js").Coordinate} coordinate Coordinate.
+ * @param {number} resolution Resolution.
+ * @return {string|undefined} GetFeatureInfo URL.
+ * @api
+ */
+export function getFeatureInfoUrl(options, coordinate, resolution) {
+  if (options.url === undefined) {
+    return undefined;
+  }
+
+  const projectionObj = getProjection(options.projection || 'EPSG:3857');
+
+  const extent = getForViewAndSize(
+    coordinate,
+    resolution,
+    0,
+    GETFEATUREINFO_IMAGE_SIZE
+  );
+
+  const baseParams = {
+    'QUERY_LAYERS': options.params['LAYERS'],
+    'INFO_FORMAT': 'application/json',
+  };
+  Object.assign(
+    baseParams,
+    getRequestParams(options.params, 'GetFeatureInfo'),
+    options.params
+  );
+
+  const x = floor((coordinate[0] - extent[0]) / resolution, DECIMALS);
+  const y = floor((extent[3] - coordinate[1]) / resolution, DECIMALS);
+  const v13 = compareVersions(baseParams['VERSION'], '1.3') >= 0;
+  baseParams[v13 ? 'I' : 'X'] = x;
+  baseParams[v13 ? 'J' : 'Y'] = y;
+
+  return getRequestUrl(
+    options.url,
+    extent,
+    GETFEATUREINFO_IMAGE_SIZE,
+    projectionObj,
+    baseParams
+  );
+}
+
+/**
+ * Get the GetLegendGraphic URL, optionally optimized for the passed resolution and possibly
+ * including any passed specific parameters. Returns `undefined` if the GetLegendGraphic URL
+ * cannot be constructed.
+ *
+ * @param {LoaderOptions} options Options passed the `createWMSLoader()` function.
+ * @param {number} [resolution] Resolution. If not provided, `SCALE` will not be calculated and
+ * included in URL.
+ * @return {string|undefined} GetLegendGraphic URL.
+ * @api
+ */
+export function getLegendUrl(options, resolution) {
+  if (options.url === undefined) {
+    return undefined;
+  }
+
+  const baseParams = {
+    'SERVICE': 'WMS',
+    'VERSION': DEFAULT_VERSION,
+    'REQUEST': 'GetLegendGraphic',
+    'FORMAT': 'image/png',
+  };
+
+  if (options.params === undefined || options.params['LAYER'] === undefined) {
+    const layers = options.params.LAYERS;
+    const isSingleLayer = !Array.isArray(layers) || layers.length === 1;
+    if (!isSingleLayer) {
+      return undefined;
+    }
+    baseParams['LAYER'] = layers;
+  }
+
+  if (resolution !== undefined) {
+    const mpu =
+      getProjection(options.projection || 'EPSG:3857').getMetersPerUnit() || 1;
+    const pixelSize = 0.00028;
+    baseParams['SCALE'] = (resolution * mpu) / pixelSize;
+  }
+
+  Object.assign(baseParams, options.params);
+
+  return appendParams(options.url, baseParams);
 }
